Refactor `dbus` module
- Created a new module to store functions that register the MPRIS interfaces
- Removed a few unwraps
- Fixed a few clippy lints
