+++ src/platform/mpris/dbus/mod.rs
@@ -1,17 +1,16 @@
+mod interfaces;
+
 use std::collections::HashMap;
 use std::convert::From;
-use std::convert::{TryFrom, TryInto};
+use std::convert::TryInto;
 use std::sync::{mpsc, Arc, Mutex};
 use std::thread::{self, JoinHandle};
 use std::time::Duration;
 
-use crate::{
-    MediaControlEvent, MediaMetadata, MediaPlayback, MediaPosition, PlatformConfig, SeekDirection,
-};
+use crate::{MediaControlEvent, MediaMetadata, MediaPlayback, PlatformConfig};
 
 /// A platform-specific error.
-#[derive(Debug)]
-pub struct Error;
+pub type Error = dbus::Error;
 
 /// A handle to OS media controls.
 pub struct MediaControls {
@@ -112,6 +111,7 @@ impl From<MediaMetadata<'_>> for OwnedMetadata {
             artist: other.artist.map(|s| s.to_string()),
             album: other.album.map(|s| s.to_string()),
             cover_url: other.cover_url.map(|s| s.to_string()),
+            // TODO: This should probably not have an unwrap
             duration: other.duration.map(|d| d.as_micros().try_into().unwrap()),
         }
     }
@@ -196,8 +196,6 @@ use dbus::channel::{MatchingReceiver, Sender};
 use dbus::ffidisp::stdintf::org_freedesktop_dbus::PropertiesPropertiesChanged;
 use dbus::message::SignalArgs;
 use dbus::Path;
-use dbus_crossroads::Crossroads;
-use dbus_crossroads::IfaceBuilder;
 
 fn run_service<F>(
     dbus_name: String,
@@ -208,217 +206,20 @@ fn run_service<F>(
 where
     F: Fn(MediaControlEvent) + Send + 'static,
 {
+    let c = Connection::new_session()?;
+    let name = format!("org.mpris.MediaPlayer2.{}", dbus_name);
+    c.request_name(name, false, true, false)?;
+
+    let state = Arc::new(Mutex::new(ServiceState {
+        metadata: Default::default(),
+        metadata_dict: create_metadata_dict(&Default::default()),
+        playback_status: MediaPlayback::Stopped,
+        volume: 1.0,
+    }));
     let event_handler = Arc::new(Mutex::new(event_handler));
     let seeked_signal = Arc::new(Mutex::new(None));
 
-    let c = Connection::new_session()?;
-    c.request_name(
-        format!("org.mpris.MediaPlayer2.{}", dbus_name),
-        false,
-        true,
-        false,
-    )?;
-
-    let mut cr = Crossroads::new();
-
-    let app_interface = cr.register("org.mpris.MediaPlayer2", {
-        let event_handler = event_handler.clone();
-
-        move |b| {
-            b.property("Identity")
-                .get(move |_, _| Ok(friendly_name.clone()));
-
-            register_method(b, &event_handler, "Raise", MediaControlEvent::Raise);
-            register_method(b, &event_handler, "Quit", MediaControlEvent::Quit);
-
-            // TODO: allow user to set these properties
-            b.property("CanQuit")
-                .get(|_, _| Ok(true))
-                .emits_changed_true();
-            b.property("CanRaise")
-                .get(|_, _| Ok(true))
-                .emits_changed_true();
-            b.property("HasTracklist")
-                .get(|_, _| Ok(false))
-                .emits_changed_true();
-            b.property("SupportedUriSchemes")
-                .get(move |_, _| Ok(&[] as &[String]))
-                .emits_changed_true();
-            b.property("SupportedMimeTypes")
-                .get(move |_, _| Ok(&[] as &[String]))
-                .emits_changed_true();
-        }
-    });
-
-    let mut state = ServiceState {
-        metadata: Default::default(),
-        metadata_dict: HashMap::new(),
-        playback_status: MediaPlayback::Stopped,
-        volume: 1.0,
-    };
-
-    state.set_metadata(Default::default());
-
-    let state = Arc::new(Mutex::new(state));
-
-    let player_interface = cr.register("org.mpris.MediaPlayer2.Player", |b| {
-        register_method(b, &event_handler, "Next", MediaControlEvent::Next);
-        register_method(b, &event_handler, "Previous", MediaControlEvent::Previous);
-        register_method(b, &event_handler, "Pause", MediaControlEvent::Pause);
-        register_method(b, &event_handler, "PlayPause", MediaControlEvent::Toggle);
-        register_method(b, &event_handler, "Stop", MediaControlEvent::Stop);
-        register_method(b, &event_handler, "Play", MediaControlEvent::Play);
-
-        b.method("Seek", ("Offset",), (), {
-            let event_handler = event_handler.clone();
-
-            move |ctx, _, (offset,): (i64,)| {
-                let abs_offset = offset.abs() as u64;
-                let direction = if offset > 0 {
-                    SeekDirection::Forward
-                } else {
-                    SeekDirection::Backward
-                };
-
-                (event_handler.lock().unwrap())(MediaControlEvent::SeekBy(
-                    direction,
-                    Duration::from_micros(abs_offset),
-                ));
-                ctx.push_msg(ctx.make_signal("Seeked", ()));
-                Ok(())
-            }
-        });
-
-        b.method("SetPosition", ("TrackId", "Position"), (), {
-            let state = state.clone();
-            let event_handler = event_handler.clone();
-
-            move |_, _, (_trackid, position): (Path, i64)| {
-                let state = state.lock().unwrap();
-
-                // According to the MPRIS specification:
-
-                // TODO: If the TrackId argument is not the same as the current
-                // trackid, the call is ignored as stale.
-                // (Maybe it should be optional?)
-
-                if let Some(duration) = state.metadata.duration {
-                    // If the Position argument is greater than the track length, do nothing.
-                    if position > duration {
-                        return Ok(());
-                    }
-                }
-
-                // If the Position argument is less than 0, do nothing.
-                if let Ok(position) = u64::try_from(position) {
-                    let position = Duration::from_micros(position);
-
-                    (event_handler.lock().unwrap())(MediaControlEvent::SetPosition(MediaPosition(
-                        position,
-                    )));
-                }
-                Ok(())
-            }
-        });
-
-        b.method("OpenUri", ("Uri",), (), {
-            let event_handler = event_handler.clone();
-
-            move |_, _, (uri,): (String,)| {
-                (event_handler.lock().unwrap())(MediaControlEvent::OpenUri(uri));
-                Ok(())
-            }
-        });
-
-        *seeked_signal.lock().unwrap() = Some(b.signal::<(String,), _>("Seeked", ("x",)).msg_fn());
-
-        b.property("PlaybackStatus")
-            .get({
-                let state = state.clone();
-                move |_, _| {
-                    let state = state.lock().unwrap();
-                    Ok(state.get_playback_status().to_string())
-                }
-            })
-            .emits_changed_true();
-
-        b.property("Rate").get(|_, _| Ok(1.0)).emits_changed_true();
-
-        b.property("Metadata")
-            .get({
-                let state = state.clone();
-                move |_, _| Ok(create_metadata_dict(&state.lock().unwrap().metadata))
-            })
-            .emits_changed_true();
-
-        b.property("Volume")
-            .get({
-                let state = state.clone();
-                move |_, _| {
-                    let state = state.lock().unwrap();
-                    Ok(state.volume)
-                }
-            })
-            .set({
-                let event_handler = event_handler.clone();
-                move |_, _, volume: f64| {
-                    (event_handler.lock().unwrap())(MediaControlEvent::SetVolume(volume));
-                    Ok(Some(volume))
-                }
-            })
-            .emits_changed_true();
-
-        b.property("Position").get({
-            let state = state.clone();
-            move |_, _| {
-                let state = state.lock().unwrap();
-                let progress: i64 = match state.playback_status {
-                    MediaPlayback::Playing {
-                        progress: Some(progress),
-                    }
-                    | MediaPlayback::Paused {
-                        progress: Some(progress),
-                    } => progress.0.as_micros(),
-                    _ => 0,
-                }
-                .try_into()
-                .unwrap();
-                Ok(progress)
-            }
-        });
-
-        b.property("MinimumRate")
-            .get(|_, _| Ok(1.0))
-            .emits_changed_true();
-        b.property("MaximumRate")
-            .get(|_, _| Ok(1.0))
-            .emits_changed_true();
-
-        b.property("CanGoNext")
-            .get(|_, _| Ok(true))
-            .emits_changed_true();
-        b.property("CanGoPrevious")
-            .get(|_, _| Ok(true))
-            .emits_changed_true();
-        b.property("CanPlay")
-            .get(|_, _| Ok(true))
-            .emits_changed_true();
-        b.property("CanPause")
-            .get(|_, _| Ok(true))
-            .emits_changed_true();
-        b.property("CanSeek")
-            .get(|_, _| Ok(true))
-            .emits_changed_true();
-        b.property("CanControl")
-            .get(|_, _| Ok(true))
-            .emits_changed_true();
-    });
-
-    cr.insert(
-        "/org/mpris/MediaPlayer2",
-        &[app_interface, player_interface],
-        (),
-    );
+    let mut cr = interfaces::register_methods(&state, &event_handler, friendly_name, seeked_signal);
 
     c.start_receive(
         dbus::message::MatchRule::new_method_call(),
@@ -470,26 +271,10 @@ where
             c.send(
                 properties_changed.to_emit_message(&Path::new("/org/mpris/MediaPlayer2").unwrap()),
             )
-            .unwrap();
+            .ok();
         }
         c.process(Duration::from_millis(1000))?;
     }
 
     Ok(())
-}
-
-fn register_method<F>(
-    b: &mut IfaceBuilder<()>,
-    event_handler: &Arc<Mutex<F>>,
-    name: &'static str,
-    event: MediaControlEvent,
-) where
-    F: Fn(MediaControlEvent) + Send + 'static,
-{
-    let event_handler = event_handler.clone();
-
-    b.method(name, (), (), move |_, _, _: ()| {
-        (event_handler.lock().unwrap())(event.clone());
-        Ok(())
-    });
 }
